+++ plutarch-orphanage/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+* Orphan instances for V3 ledger types
 
 ## 1.0.0 -- 20-06-2024